+++ api/src/main/scala/hmda/api/HmdaApi.scala
@@ -9,11 +9,11 @@ import akka.stream.ActorMaterializer
 import akka.util.Timeout
 import com.typesafe.config.ConfigFactory
 import hmda.api.http._
+import hmda.api.http.institutions.InstitutionsHttpApi
 import hmda.persistence.HmdaSupervisor
 import hmda.persistence.HmdaSupervisor.FindActorByName
 import hmda.persistence.demo.DemoData
 import hmda.persistence.institutions.InstitutionPersistence
-import hmda.persistence.institutions.InstitutionPersistence._
 import hmda.persistence.processing.{ LocalHmdaEventProcessor, SingleLarValidation }
 
 import scala.concurrent.duration._
