+++ api/src/main/scala/hmda/api/http/institutions/InstitutionsHttpApi.scala
@@ -1,0 +1,40 @@
+package hmda.api.http.institutions
+
+import akka.actor.ActorSystem
+import akka.event.LoggingAdapter
+import akka.http.scaladsl.server.Directives._
+import akka.stream.ActorMaterializer
+import akka.util.Timeout
+import hmda.api.http.HmdaCustomDirectives
+import hmda.api.protocol.processing.{ ApiErrorProtocol, InstitutionProtocol }
+
+trait InstitutionsHttpApi
+    extends InstitutionProtocol
+    with InstitutionPaths
+    with FilingPaths
+    with SubmissionPaths
+    with UploadPaths
+    with ApiErrorProtocol
+    with HmdaCustomDirectives {
+
+  implicit val system: ActorSystem
+  implicit val materializer: ActorMaterializer
+  val log: LoggingAdapter
+
+  implicit val timeout: Timeout
+
+  val institutionsRoutes =
+    headerAuthorize {
+      institutionsPath ~
+        pathPrefix("institutions" / Segment) { instId =>
+          institutionAuthorize(instId) {
+            institutionByIdPath(instId) ~
+              institutionSummaryPath(instId) ~
+              filingByPeriodPath(instId) ~
+              submissionPath(instId) ~
+              submissionLatestPath(instId) ~
+              uploadPath(instId)
+          }
+        }
+    }
+}
